Time: 5 ms (89.98%), Space: 48.6 MB (7.66%) - LeetHub

+++ 3194-minimum-average-of-smallest-and-largest-elements/3194-minimum-average-of-smallest-and-largest-elements.java
@@ -1,9 +1,9 @@
 class Solution {
     public double minimumAverage(int[] nums) {
-         double d =Integer.MAX_VALUE;
+         double d =Double.MAX_VALUE;
         Arrays.sort(nums);
         for(int i=0;i<nums.length/2;i++){
-            d=Math.min(d,((double)(nums[i]+nums[nums.length-i-1])/2));
+            d=Math.min(d,(((double)nums[i]+nums[nums.length-i-1])/2));
         }
         return d;
     }
